+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Doppio Changelog
 
+## v0.6.0
+
+*unreleased*
+
 ## v0.5.0
+
+*2021-04-24*
 
 ### Added